+++ exchangelib/autodiscover/discovery.py
@@ -1,4 +1,3 @@
-import abc
 import logging
 from urllib.parse import urlparse
 
@@ -6,11 +5,13 @@ import dns.name
 import dns.resolver
 from cached_property import threaded_cached_property
 
-from ...errors import AutoDiscoverCircularRedirect, AutoDiscoverFailed, TransportError
-from ...protocol import FailFast
-from ...util import DummyResponse, get_domain, get_redirect_url
-from ..cache import autodiscover_cache
-from ..protocol import AutodiscoverProtocol
+from ..configuration import Configuration
+from ..errors import AutoDiscoverCircularRedirect, AutoDiscoverFailed, RedirectError, TransportError
+from ..protocol import FailFast, Protocol
+from ..transport import get_unauthenticated_autodiscover_response
+from ..util import CONNECTION_ERRORS, DummyResponse, get_domain, get_redirect_url
+from .cache import autodiscover_cache
+from .protocol import AutodiscoverProtocol
 
 log = logging.getLogger(__name__)
 
@@ -35,7 +36,14 @@ class SrvRecord:
         return all(getattr(self, k) == getattr(other, k) for k in self.__dict__)
 
 
-class BaseAutodiscovery(metaclass=abc.ABCMeta):
+def discover(email, credentials=None, auth_type=None, retry_policy=None):
+    ad_response, protocol = Autodiscovery(email=email, credentials=credentials).discover()
+    protocol.config.auth_typ = auth_type
+    protocol.config.retry_policy = retry_policy
+    return ad_response, protocol
+
+
+class Autodiscovery:
     """Autodiscover is a Microsoft protocol for automatically getting the endpoint of the Exchange server and other
     connection-related settings holding the email address using only the email address, and username and password of the
     user.
@@ -72,7 +80,7 @@ class BaseAutodiscovery(metaclass=abc.ABCMeta):
         "timeout": AutodiscoverProtocol.TIMEOUT / 2.5,  # Timeout for query to a single nameserver
     }
     DNS_RESOLVER_LIFETIME = AutodiscoverProtocol.TIMEOUT  # Total timeout for a query in case of multiple nameservers
-    URL_PATH = None
+    URL_PATH = "autodiscover/autodiscover.svc"
 
     def __init__(self, email, credentials=None):
         """
@@ -146,13 +154,27 @@ class BaseAutodiscovery(metaclass=abc.ABCMeta):
             setattr(resolver, k, v)
         return resolver
 
-    @abc.abstractmethod
     def _build_response(self, ad_response):
-        pass
+        if not ad_response.autodiscover_smtp_address:
+            # Autodiscover does not always return an email address. In that case, the requesting email should be used
+            ad_response.autodiscover_smtp_address = self.email
 
-    @abc.abstractmethod
+        protocol = Protocol(
+            config=Configuration(
+                service_endpoint=ad_response.ews_url,
+                credentials=self.credentials,
+                version=ad_response.version,
+                # TODO: Detect EWS service auth type somehow
+            )
+        )
+        return ad_response, protocol
+
     def _quick(self, protocol):
-        pass
+        try:
+            user_response = protocol.get_user_settings(user=self.email)
+        except TransportError as e:
+            raise AutoDiscoverFailed(f"Response error: {e}")
+        return self._step_5(ad=user_response)
 
     def _redirect_url_is_valid(self, url):
         """Three separate responses can be “Redirect responses”:
@@ -190,13 +212,66 @@ class BaseAutodiscovery(metaclass=abc.ABCMeta):
         self._redirect_count += 1
         return True
 
-    @abc.abstractmethod
     def _get_unauthenticated_response(self, url, method="post"):
-        pass
+        """Get response from server using the given HTTP method
 
-    @abc.abstractmethod
+        :param url:
+        :return:
+        """
+        # We are connecting to untrusted servers here, so take necessary precautions.
+        self._ensure_valid_hostname(url)
+
+        protocol = AutodiscoverProtocol(
+            config=Configuration(
+                service_endpoint=url,
+                retry_policy=self.INITIAL_RETRY_POLICY,
+            )
+        )
+        return None, get_unauthenticated_autodiscover_response(protocol=protocol, method=method)
+
     def _attempt_response(self, url):
-        pass
+        """Return an (is_valid_response, response) tuple.
+
+        :param url:
+        :return:
+        """
+        self._urls_visited.append(url.lower())
+        log.debug("Attempting to get a valid response from %s", url)
+
+        try:
+            self._ensure_valid_hostname(url)
+        except TransportError:
+            return False, None
+
+        protocol = AutodiscoverProtocol(
+            config=Configuration(
+                service_endpoint=url,
+                credentials=self.credentials,
+                retry_policy=self.INITIAL_RETRY_POLICY,
+            )
+        )
+        try:
+            user_response = protocol.get_user_settings(user=self.email)
+        except RedirectError as e:
+            if self._redirect_url_is_valid(url=e.url):
+                # The protocol does not specify this explicitly, but by looking at how testconnectivity.microsoft.com
+                # works, it seems that we should follow this URL now and try to get a valid response.
+                return self._attempt_response(url=e.url)
+            log.debug("Invalid redirect URL: %s", e.url)
+            return False, None
+        except TransportError as e:
+            log.debug("Failed to get a response: %s", e)
+            return False, None
+        except CONNECTION_ERRORS as e:
+            log.debug("Failed to get a response: %s", e)
+            return False, None
+
+        # We got a valid response. Unless this is a URL redirect response, we cache the result
+        if not user_response.redirect_url:
+            cache_key = self._cache_key
+            log.debug("Adding cache entry for key %s: %s", cache_key, protocol.service_endpoint)
+            autodiscover_cache[cache_key] = protocol
+        return True, user_response
 
     def _ensure_valid_hostname(self, url):
         hostname = urlparse(url).netloc
@@ -364,10 +439,6 @@ class BaseAutodiscovery(metaclass=abc.ABCMeta):
         log.info("Step 5: Checking response")
         # This is not explicit in the protocol, but let's raise any errors here
         ad.raise_errors()
-
-        if hasattr(ad, "response"):
-            # Hack for PoxAutodiscover
-            ad = ad.response
 
         if ad.redirect_url:
             log.debug("Got a redirect URL: %s", ad.redirect_url)